+++ src/main/scala/components/alucontrol.scala
@@ -1,6 +1,6 @@
 // This file contains ALU control logic.
 
-package dinocpu
+package dinocpu.components
 
 import chisel3._
 import chisel3.util._
