+++ src/main/scala/components/coreio.scala
@@ -1,6 +1,6 @@
 // The IO between the core and the rest of the system
 
-package dinocpu
+package dinocpu.components
 
 import chisel3._
 import dinocpu.memory._
