+++ src/main/scala/pipelined/cpu-bp.scala
@@ -3,9 +3,10 @@
 
 package dinocpu.pipelined
 
-import dinocpu._
 import chisel3._
 import chisel3.util._
+import dinocpu._
+import dinocpu.components._
 
 /**
  * The main CPU definition that hooks up all of the other components.
