+++ src/main/scala/pipelined/cpu.scala
@@ -5,6 +5,7 @@ package dinocpu.pipelined
 import chisel3._
 import chisel3.util._
 import dinocpu._
+import dinocpu.components._
 
 /**
  * The main CPU definition that hooks up all of the other components.
