+++ src/test/scala/components/ALUControlUnitTest.scala
@@ -4,7 +4,7 @@ package dinocpu.test.components
 
 import chisel3._
 import chisel3.iotesters.{ChiselFlatSpec, Driver, PeekPokeTester}
-import dinocpu._
+import dinocpu.components.ALUControl
 
 class ALUControlUnitTester(c: ALUControl) extends PeekPokeTester(c) {
   private val ctl = c
