+++ src/test/scala/components/ALUUnitTest.scala
@@ -4,7 +4,7 @@ package dinocpu.test.components
 
 import chisel3._
 import chisel3.iotesters.{ChiselFlatSpec, Driver, PeekPokeTester}
-import dinocpu._
+import dinocpu.components.ALU
 
 
 class ALURandomUnitTester(c: ALU) extends PeekPokeTester(c) {
