+++ src/test/scala/components/BranchControlUnitTest.scala
@@ -4,7 +4,7 @@ package dinocpu.test.components
 
 import chisel3._
 import chisel3.iotesters.{ChiselFlatSpec, Driver, PeekPokeTester}
-import dinocpu._
+import dinocpu.components._
 
 
 class BranchControlUnitTester(c: BranchControl, branch: Boolean = true) extends PeekPokeTester(c) {
