+++ src/test/scala/grading/Lab1Tests.scala
@@ -3,7 +3,9 @@
 package dinocpu.test.grader
 
 import dinocpu._
+import dinocpu.components._
 import dinocpu.test._
+import dinocpu.test.components._
 import chisel3.iotesters.Driver
 
 import com.gradescope.jh61b.grader.{GradedTest,GradedTestRunnerJSON}
